Fully functional site with jsp for backend
This is a fully functional site, created using html, css, javascript,
java server pages, relational database , jquery, lightbox for picture
slide show,
and google maps api.

The concept of the site was to create an online platform in which, a
person
could register his house, for sale or rent, while others could show
interest
to it, by emailing to that person, or sending him an inbox message.

A person who wishes to rent or sell his/her house, should first register
to
our database as a seller (to sell) or landlord (to rent) or both.
After that, he can write the location of the house on the register a
house
platform, and it automatically locates it on the google map, or he can
just
click on the map to add it as well.

Every house has its own profile page, where you can see info and
pictures about
it.If you are the house owner you can configure this info, and
upload/delete
pictures.

Additionally, when a seller/landlord logs in, the map displays clickable
icons
for the houses he/she has already registered to our database.

Now, a person who wishes to buy, or rent a house, he or she, can enter
in
a search platform his/hers personal preferences and search for it.
In order to contact the seller/landlord, one has to register to our
platform
firstly.

The search is being made with two differnet algorithms.
Simple Additive Weight (SAW) algorithm
https://en.wikipedia.org/wiki/Weighted_sum_model
and
TOPSIS algorithm https://en.wikipedia.org/wiki/TOPSIS